requirements_CondaYes.txt
absl-py==0.3.0
alabaster==0.7.11
asn1crypto==0.24.0
astor==0.7.1
Babel==2.6.0
backcall==0.1.0
bleach==2.1.3
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cloudpickle==0.5.3
cryptography==2.3
cycler==0.10.0
Cython==0.28.4
dask==0.18.2
decorator==4.3.0
docutils==0.14
entrypoints==0.2.3
gast==0.2.0
graphviz==0.8.4
grpcio==1.13.0
h5py==2.8.0
html5lib==1.0.1
idna==2.7
imagesize==1.0.0
ipykernel==4.8.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.3.1
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.4
Keras-Preprocessing==1.0.2
kiwisolver==1.0.1
leveldb==0.194
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
ndg-httpsclient==0.5.1
networkx==2.1
nose==1.3.7
notebook==5.6.0
numpy==1.13.3
packaging==17.1
pandas==0.23.3
pandocfilters==1.4.2
parso==0.3.1
path.py==11.0.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
pip==18.0
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.0
ptyprocess==0.6.0
pyasn1==0.4.4
pycparser==2.18
pydot==1.2.4
pydot-ng==1.0.0
Pygments==2.2.0
pygobject==3.26.1
pyOpenSSL==18.0.0
pyparsing==2.2.0
python-apt==1.6.2
python-dateutil==2.7.3
python-gflags==3.1.2
pytz==2018.5
PyWavelets==0.5.2
PyYAML==3.13
pyzmq==17.1.0
qtconsole==4.3.1
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
Send2Trash==1.5.0
setuptools==39.1.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.6
sphinxcontrib-websupport==1.1.0
tensorboard==1.9.0
tensorflow-gpu==1.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==5.1
traitlets==4.3.2
unattended-upgrades==0.1
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
wheel==0.31.1
widgetsnbextension==3.3.1
zmq==0.0.0
git+git://github.com/slundberg/shap.git@41873154561cfa059fa8ed1fc8dec1095b1de9d1#egg=shap
